+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.15.1] - 2026-06-03
+
+### Bug Fix
+
+1. Node reference validation while import
+   - Not to validate node references before inserting to avoid unnecessary performance hit
+
 ## [7.15.0] - 2026-06-03
 
 ### ✨ New Features